+++ data/src/main/java/com/tahn/assignment/remote/GithubRemoteDataSource.kt
@@ -3,5 +3,8 @@ package com.tahn.assignment.remote
 import com.tahn.assignment.model.remote.GithubUserResponse
 
 interface GithubRemoteDataSource {
-    suspend fun fetchUsers(perPage: Int, since: Long): List<GithubUserResponse>
+    suspend fun fetchUsers(
+        perPage: Long,
+        since: Int,
+    ): List<GithubUserResponse>
 }
